# app/static/swagger.yaml
swagger: '2.0'
info:
  title: 'CryptoTemple API'
  description: 'Cryptocurrency trading simulation platform API'
  version: '1.0.0'
host: 'localhost:5000'
schemes:
  - 'http'
consumes:
  - 'application/json'
produces:
  - 'application/json'
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: 'JWT token for authentication (format: Bearer <token>)'
paths:
...
  /trade/{user_id}/order:
    post:
      tags:
        - Trading
      summary: "Place new order"
      description: "Create a new buy or sell order (rate limited to 20 per minute)"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "user_id"
          type: integer
          required: true
          description: "User ID"
        - in: "body"
          name: "body"
          description: "Order details"
          required: true
          schema:
            $ref: "#/definitions/PlaceOrderRequest"
      responses:
        201:
          description: "Order placed successfully"
          schema:
            $ref: "#/definitions/OrderResponse"
        400:
          description: "Invalid order or insufficient balance"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Unauthorized access"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Cryptocurrency not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /trade/{user_id}/orders:
    get:
      tags:
        - Trading
      summary: "Get user's orders"
      description: "Retrieve paginated list of user's orders with optional filtering"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "user_id"
          type: integer
          required: true
          description: "User ID"
        - in: "query"
          name: "page"
          type: integer
          required: false
          default: 1
          description: "Page number"
        - in: "query"
          name: "per_page"
          type: integer
          required: false
          default: 10
          description: "Items per page"
        - in: "query"
          name: "status"
          type: string
          required: false
          enum: ["pending", "completed", "cancelled"]
          description: "Filter by order status"
      responses:
        200:
          description: "Orders retrieved successfully"
          schema:
            $ref: "#/definitions/PaginatedOrdersResponse"
        403:
          description: "Unauthorized access"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "User not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /trade/{user_id}/order/{order_id}:
    get:
      tags:
        - Trading
      summary: "Get specific order"
      description: "Retrieve details of a specific order"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "user_id"
          type: integer
          required: true
          description: "User ID"
        - in: "path"
          name: "order_id"
          type: integer
          required: true
          description: "Order ID"
      responses:
        200:
          description: "Order details retrieved"
          schema:
            $ref: "#/definitions/OrderDetailResponse"
        403:
          description: "Unauthorized access"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Order not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /trade/{user_id}/order/{order_id}/execute:
    post:
      tags:
        - Trading
      summary: "Execute pending order"
      description: "Execute a pending order and update portfolio holdings"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "user_id"
          type: integer
          required: true
          description: "User ID"
        - in: "path"
          name: "order_id"
          type: integer
          required: true
          description: "Order ID to execute"
      responses:
        200:
          description: "Order executed successfully"
          schema:
            $ref: "#/definitions/ExecuteOrderResponse"
        400:
          description: "Order cannot be executed (invalid status or insufficient balance)"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Unauthorized access"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Order not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /trade/{user_id}/order/{order_id}/cancel:
    post:
      tags:
        - Trading
      summary: "Cancel order"
      description: "Cancel a pending order"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "user_id"
          type: integer
          required: true
          description: "User ID"
        - in: "path"
          name: "order_id"
          type: integer
          required: true
          description: "Order ID to cancel"
      responses:
        200:
          description: "Order cancelled successfully"
          schema:
            $ref: "#/definitions/MessageResponse"
        400:
          description: "Order cannot be cancelled (not pending)"
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: "Unauthorized access"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Order not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

# ============================================================================
# PORTFOLIO ENDPOINTS
# ============================================================================
...
definitions:
...
  PlaceOrderRequest:
    type: object
    required:
      - crypto_id
      - order_type
      - quantity
      - price
    properties:
      crypto_id:
        type: integer
        example: 1
      order_type:
        type: string
        enum: ["buy", "sell"]
        example: "buy"
      quantity:
        type: number
        format: float
        example: 0.5
      price:
        type: number
        format: float
        example: 45000.00
  OrderResponse:
    type: object
    properties:
      message:
        type: string
      order:
        $ref: "#/definitions/OrderData"
      new_cash_balance:
        type: number
        format: float
  OrderData:
    type: object
    properties:
      id:
        type: integer
      type:
        type: string
        enum: ["buy", "sell"]
      symbol:
        type: string
      quantity:
        type: number
        format: float
      price:
        type: number
        format: float
      total_value:
        type: number
        format: float
      status:
        type: string
        enum: ["pending", "completed", "cancelled"]
      created_at:
        type: string
        format: date-time
  PaginatedOrdersResponse:
    type: object
    properties:
      orders:
        type: array
        items:
          $ref: "#/definitions/OrderDetailResponse"
      total:
        type: integer
      page:
        type: integer
      per_page:
        type: integer
  OrderDetailResponse:
    type: object
    properties:
      id:
        type: integer
      crypto_id:
        type: integer
      symbol:
        type: string
      type:
        type: string
      quantity:
        type: number
        format: float
      price:
        type: number
        format: float
      total_value:
        type: number
        format: float
      status:
        type: string
      created_at:
        type: string
        format: date-time
      executed_at:
        type: string
        format: date-time
        nullable: true
  ExecuteOrderResponse:
    type: object
    properties:
      message:
        type: string
      order:
        type: object
        properties:
          id:
            type: integer
          status:
            type: string
          executed_at:
            type: string
            format: date-time

  # Portfolio Definitions
...
  ErrorResponse:
    type: object
    properties:
      message:
        type: string
        example: "Error message"
  MessageResponse:
    type: object
    properties:
      message:
        type: string
        example: "Success"